Extract Provider Only Iframe Initializer for separate class
